refactor attach_method_relationship
